v4.0-0 (UNSTABLE)
Up to v3.9 the whole project was named "Ansible Engine (AE)".
Unfortunately this conflicted with RedHat's product name: ansible engine
(https://www.ansible.com/products/engine).
That's why the whole project has been renamed in v4 to
"Playbooks-On-Rails for Splunk (PORS)".

this commit includes all changes in v3.5 - 3.9.
v4 is not just a re-branding it completely re-order the whole folder
structure and a lot of clean-ups (besides several fixes and some new
features ofc).

Main reason for this is easing up the parallel operation with ansible
tower / AWX and most important it paves the way for more frequent
git commits. Atm this is tricky due to the nature of how things are set up
and so the PORS code will be more independent from your personal
configuration.

Another important change is switching the license to AGPL.

The current stage is 100% unstable so not working at all as all playbooks
need to be adapted to new paths and such. Wait until this code reaches
the stable branch!