gcp: completely re-worked instance-deploy
The GCP ansible modules do not offer to attach disks to an already
running instance. There is an old module 'gc_pd' but it is
superseeded by gcp_compute_instance. The latter does not offer an
'attach-disk' key (as in gcloud SDK) so atm (2021, Jan) there is no
way to do this task via the gcp ansible modules.

For not loosing flexibility by using hard-coded disk mappings (i.e.
hard-coded within ansible + pors) I switched to use a jinja template
instead. This way one can configure everything in a hardware profile
as before but nothing needs to be hard-coded anywhere.

Briefly check:
EXAMPLES/vm_recipes/defaults/pors_hardware-config_common.config
for a full explanation of VMDISK variables, when they need to be set
and how.

Signed-off-by: TF <mail@sedi.one>