vars:roles:git: add default branch override
this allows to specify an override for the auto-detected
default branch of git repo's in:

inventories/<env>/group_vars/all/git.yml

it has been implemented in a way that you can change or remove the
override at any time without the need to re-link apps.

HINT: this is a global setting which will affect ALL repo's of an
environment (staging,production,...).

This separation allows having different default branches depending on the
environment you are in.

Signed-off-by: TF <mail@sedi.one>