roles: enable global defaults for all roles
this changes the whole way of how variables are set in PORS.
before it was necessary to ensure that each environment have all
variables in inventories/group_vars/all .
this was not just failure prone but also become problematic when adding
or changing these global variables.

instead PORS will now make use of roles/defaults which contain all
default settings. a hack was needed to make this available for all roles
by once but it is worth it.

keep in mind that the precedence order in ansible is:

roles/defaults/main.yml
inventories/group_vars/all
.. etc
(Details: https://docs.ansible.com/ansible/latest/user_guide/playbooks_variables.html#variable-precedence-where-should-i-put-a-variable)

There are still variable ymls in inventories which need to be moved but
bc of the default hash behavior variables get replaced and not merged.
Unfortunately for history reasons there are some variables written as
dict and are not part of the new defaults approach (yet).